move validate_pedigree to utils
